renamed test base class to contain test_ prefix in order for pytest to work correctly with assert_stdout helper method
